# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for North Texas Commercial Roofing website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactInputServiceType = typeof ContactInputServiceType[keyof typeof ContactInputServiceType];


export const ContactInputServiceType = {
  'roof-repair': 'roof-repair',
  'roof-replacement': 'roof-replacement',
  inspection: 'inspection',
  maintenance: 'maintenance',
  'storm-damage': 'storm-damage',
  'emergency-leak': 'emergency-leak',
  coatings: 'coatings',
  'flat-roofing': 'flat-roofing',
  'metal-roofing': 'metal-roofing',
  'tpo-epdm': 'tpo-epdm',
  other: 'other',
} as const;

/**
 * @nullable
 */
export type ContactInputPreferredContact = typeof ContactInputPreferredContact[keyof typeof ContactInputPreferredContact] | null;


export const ContactInputPreferredContact = {
  phone: 'phone',
  email: 'email',
} as const;

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  company?: string | null;
  serviceType: ContactInputServiceType;
  /** @nullable */
  propertyType?: string | null;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  preferredContact?: ContactInputPreferredContact;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  company?: string | null;
  serviceType: string;
  /** @nullable */
  propertyType?: string | null;
  message: string;
  /** @nullable */
  preferredContact?: string | null;
  createdAt: string;
}

export type EstimatorInputSqftSource = typeof EstimatorInputSqftSource[keyof typeof EstimatorInputSqftSource];


export const EstimatorInputSqftSource = {
  solar: 'solar',
  manual: 'manual',
} as const;

export interface EstimatorInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  message?: string | null;
  /** @minLength 1 */
  address: string;
  latitude: number;
  longitude: number;
  /** @minimum 1 */
  roofSqft: number;
  sqftSource: EstimatorInputSqftSource;
  serviceType: string;
}

export interface EstimatorSubmission {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  message?: string | null;
  address: string;
  latitude: number;
  longitude: number;
  roofSqft: number;
  sqftSource: string;
  serviceType: string;
  estimatedCostUsd: number;
  pricePerSqft: number;
  createdAt: string;
}

export type RoofAreaResultRoofSegmentsItem = {
  swLat: number;
  swLng: number;
  neLat: number;
  neLng: number;
};

export interface RoofAreaResult {
  found: boolean;
  /** @nullable */
  roofAreaSqMeters?: number | null;
  /** @nullable */
  roofAreaSqft?: number | null;
  /** Bounding boxes of each detected roof segment from the Solar API, used to draw a real (non-fabricated) outline on the satellite image. Empty if no segment geometry was returned. */
  roofSegments: RoofAreaResultRoofSegmentsItem[];
}

export interface ErrorResponse {
  error: string;
}

export type GetRoofAreaParams = {
lat: number;
lng: number;
};