# lib/api-zod/src/generated/types/estimatorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for North Texas Commercial Roofing website
 * OpenAPI spec version: 0.1.0
 */
import type { EstimatorInputSqftSource } from './estimatorInputSqftSource';

export interface EstimatorInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  message?: string | null;
  /** @minLength 1 */
  address: string;
  latitude: number;
  longitude: number;
  /** @minimum 1 */
  roofSqft: number;
  sqftSource: EstimatorInputSqftSource;
  serviceType: string;
}