# lib/api-zod/src/generated/types/estimatorSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for North Texas Commercial Roofing website
 * OpenAPI spec version: 0.1.0
 */

export interface EstimatorSubmission {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  message?: string | null;
  address: string;
  latitude: number;
  longitude: number;
  roofSqft: number;
  sqftSource: string;
  serviceType: string;
  estimatedCostUsd: number;
  pricePerSqft: number;
  createdAt: Date;
}