# lib/api-zod/src/generated/types/roofAreaResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for North Texas Commercial Roofing website
 * OpenAPI spec version: 0.1.0
 */
import type { RoofAreaResultRoofSegmentsItem } from './roofAreaResultRoofSegmentsItem';

export interface RoofAreaResult {
  found: boolean;
  /** @nullable */
  roofAreaSqMeters?: number | null;
  /** @nullable */
  roofAreaSqft?: number | null;
  /** Bounding boxes of each detected roof segment from the Solar API, used to draw a real (non-fabricated) outline on the satellite image. Empty if no segment geometry was returned. */
  roofSegments: RoofAreaResultRoofSegmentsItem[];
}